Make percentages present as floats

+++ game/code/game_debug.cpp
@@ -726,8 +726,8 @@ DrawTopClocksList(debug_state *DebugState, debug_id GraphID, rectangle2 ProfileR
 
         char TextBuffer[256];
         _snprintf_s(TextBuffer, sizeof(TextBuffer),
-                    "%10ucy %3d%% %4d %s",
-                    (u32)Stats->Sum, (u32)(PC*Stats->Sum), Stats->Count, 
+                    "%10ucy %.02f%% %4d %s",
+                    (u32)Stats->Sum, (PC*Stats->Sum), Stats->Count, 
                     Element->GUID + Element->NameStartsAt);
         TextOutAt(DebugState, At, TextBuffer);
         At.y -= GetLineAdvance(DebugState);
